# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Autobot API",
        "description": "API documentation for the Autobot application",
        "contact": {
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/autobots": {
            "get": {
                "tags": [
                    "Autobots"
                ],
                "summary": "Get list of Autobots",
                "description": "Returns a paginated list of Autobots",
                "operationId": "e568a38e80296ed269a477d384fb978c",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The number of results to return per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Autobot"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/autobots/{id}": {
            "get": {
                "tags": [
                    "Autobots"
                ],
                "summary": "Get Autobot details",
                "description": "Returns details of a specific Autobot",
                "operationId": "32b0788ee3be6e44082629719c4cf7eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Autobot to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autobot"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Autobot not found"
                    }
                }
            }
        },
        "/api/autobots/{id}/posts": {
            "get": {
                "tags": [
                    "Autobots"
                ],
                "summary": "Get posts of a specific Autobot",
                "description": "Returns a paginated list of posts for a specific Autobot",
                "operationId": "cfaee44490635566c660f60d38cc98b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Autobot",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Autobot": {
                "title": "Autobot",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Autobot Name"
                    },
                    "email": {
                        "type": "string",
                        "example": "autobot@example.com"
                    },
                    "username": {
                        "type": "string",
                        "example": "autobot123"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Post": {
                "title": "Post",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "autobot_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Post Title"
                    },
                    "body": {
                        "type": "string",
                        "example": "Post content"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Autobots",
            "description": "Autobots"
        }
    ]
}